requirements.txt
absl-py==1.1.0
asgiref==3.5.2
beautifulsoup4==4.11.1
boto3==1.24.7
botocore==1.27.7
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
Django==4.0.5
django-cors-headers==3.13.0
django-extra-fields==3.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
ebird-api==3.0.6
fire==0.4.0
fonttools==4.33.3
google-auth==2.7.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
idna==3.3
importlib-metadata==4.11.4
jmespath==1.0.0
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
sahi==0.9.4
scipy==1.8.1
seaborn==0.11.2
Shapely==1.8.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminaltables==3.1.10
thop==0.1.0.post2206102148
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wikipedia==1.4.0
Wikipedia-API==0.5.4
yolov5==6.1.3
zipp==3.8.0
